changing to constants instead of hardcoding the url

+++ includes/services/FacebookService.php
@@ -10,20 +10,24 @@ use NewfoldLabs\WP\Module\Facebook\Services\FacebookHelperService;
 class FacebookService {
 
     public static function get_hiive_token(){
-        $hiive_token = HiiveConnection::get_auth_token() ? HiiveConnection::get_auth_token() :  'test6';
+        $hiive_token = HiiveConnection::get_auth_token() ? HiiveConnection::get_auth_token() :  'test2';
         return $hiive_token;
     }
 
     public static function get_token(){
-        $hiive_token = HiiveConnection::get_auth_token() ? HiiveConnection::get_auth_token() :  'test6';
-        $url = 'https://facebook-connect.bluehost.workers.dev/get/token?hiive_token='. $hiive_token;
-        $result = wp_remote_get($url, array(
+        $hiive_token = HiiveConnection::get_auth_token() ? HiiveConnection::get_auth_token() :  'test2';
+        $url = NFD_FACECBOOK_WORKER.'/get/token?hiive_token='. $hiive_token;
+        $request = wp_remote_get($url, array(
             'headers' => array(
                 'Content-Type'  => 'application/json',
                 'Accept'        => 'application/json',
             ),
         ));
-       $response = json_decode(wp_remote_retrieve_body( $result ));
+       $response = json_decode(wp_remote_retrieve_body( $request ));
+
+       if(is_wp_error($request)){
+        return array("error" => "we're unable to process the request!");
+        }
         if($response->token){
             $details = array(
                 "token" => UtilityService::encrypt_token($response->token),
@@ -36,16 +40,19 @@ class FacebookService {
     }
 
     public static function delete_token(){
-        $hiive_token = HiiveConnection::get_auth_token() ? HiiveConnection::get_auth_token() :  'test6';
-        $url = 'https://facebook-connect.bluehost.workers.dev/delete/token?hiive_token='. $hiive_token;
-        $result = wp_remote_get($url, array(
+        $hiive_token = HiiveConnection::get_auth_token() ? HiiveConnection::get_auth_token() :  'test2';
+        $url = NFD_FACECBOOK_WORKER.'/delete/token?hiive_token='. $hiive_token;
+        $request = wp_remote_get($url, array(
             'headers' => array(
             'Content-Type'  => 'application/json',
             'Accept'        => 'application/json',
         ),
-    ));
+        ));
+        if(is_wp_error($request)){
+            return array("error" => "we're unable to process the request!");
+        }
         UtilityService::deleteTokenFromCookie();
-        return $result;
+        return array("message" => "token deteled successfully!");
     }
 
     public static function get_fb_details(){
@@ -60,22 +67,27 @@ class FacebookService {
             $fb_token = FacebookService::get_token();
         }
         if(isset($fb_token) && $fb_token){     
-        $url = 'https://graph.facebook.com/me?fields=id,name,email,picture&access_token='. $fb_token;
-        $result = wp_remote_get($url, array(
+        $url = NFD_FACECBOOK_GRAPH_BASE.'/me?fields=id,name,email,picture&access_token='. $fb_token;
+        $request = wp_remote_get($url, array(
                 'headers' => array(
                 'Content-Type'  => 'application/json',
                 'Accept'        => 'application/json',
             ),
         ));
-        $response = json_decode(wp_remote_retrieve_body( $result ));
+        if(is_wp_error($request)){
+            return array("error" => "Error while trying to fetch the data from facebook!");
+        }
+        $response = json_decode(wp_remote_retrieve_body( $request ));
         if ( $response && $response->id) {
             FacebookHelperService::get_fb_posts($response, $FacebookData, $fb_token);
             FacebookHelperService::get_fb_images($response, $FacebookData, $fb_token);
             FacebookHelperService::get_fb_business($response, $FacebookData, $fb_token);
             FacebookHelperService::get_fb_business_posts($response, $FacebookData, $fb_token);
             FacebookHelperService::get_fb_business_images($response, $FacebookData, $fb_token);
-            
+        }else{
+            return array("error" => "we're unable to process the request");
         }
+
         $FacebookData->set_source("facebook");
         $FacebookData->get_Users()->set_profile($response);
        
@@ -83,7 +95,7 @@ class FacebookService {
         update_option('nfd_fb_details', $FacebookData);
         return array($FacebookData);
         }
-        return "token not found!";
+        return array("error" =>"token not found!");
     }
 }
 ?>
